PDC-77 Avance backend procesamiento de ZIP con pdfs de pago SUA

+++ app/Services/SEGURIDAD_ERP/PadronProveedores/ArchivoService.php
@@ -4,7 +4,10 @@
 namespace App\Services\SEGURIDAD_ERP\PadronProveedores;
 
 
+use FilesystemIterator;
 use Chumper\Zipper\Zipper;
+use RecursiveIteratorIterator;
+use RecursiveDirectoryIterator;
 use Illuminate\Support\Facades\Storage;
 use App\Repositories\Repository as Repository;
 use App\Models\SEGURIDAD_ERP\PadronProveedores\Archivo;
@@ -25,7 +28,7 @@ class ArchivoService
         $this->repository = new Repository($model);
     }
 
-    public function cargarArchivo_bis($data){
+    public function cargarArchivo($data){
         $directorio = $data['rfc'];
         $hash_file = hash_file('md5', $data["archivo"]);
         $archivo = $this->repository->show($data['id_archivo']);
@@ -58,14 +61,22 @@ class ArchivoService
         return $archivo;
     }
 
-    public function cargarArchivo($data){
-        // require('fpdf_merge.php');
-        // dd('panda', $data['archivo']);
+    public function cargarArchivoZIP($data){
+        $directorio = $data['rfc'];
+        if(array_key_exists('rfc_empresa', $data)){
+            $directorio = $data['rfc_empresa'] . '/' . $directorio;
+        }
+        
+        $archivo = $this->repository->show($data['id_archivo']);
+        if($archivo->usuario_registro && $archivo->usuario_registro != auth()->id()){
+            abort(403, 'No puede actualizar el archivo porque fue registrado por otro usuario.');
+        }
+        
         $paths = $this->generaDirectorios();
         $exp = explode("base64,", $data['archivo']);
-        $data = base64_decode($exp[1]);
+        $decode = base64_decode($exp[1]);
         $file = public_path($paths["path_zip"]);
-        file_put_contents($file, $data);
+        file_put_contents($file, $decode);
         $zipper = new Zipper;
         $zipper->make(public_path($paths["path_zip"]))->extractTo(public_path($paths["path_pdf"]));
         $zipper->delete();
@@ -77,23 +88,50 @@ class ArchivoService
             // dd($paths["path_pdf"]. $file);
             $pdf->addPDF($paths["path_pdf"]. $file, 'all');
         }
-        $pdf->merge('file', $paths["path_pdf"].'TEST2.pdf', 'P');
+        $pdf->merge('file', $paths["path_pdf"].'temp_pdf.pdf', 'P');
 
+        $pdf_file = fopen($paths["path_pdf"].'temp_pdf.pdf', 'r');
 
+        $hash_file = hash_file('md5', $paths["path_pdf"].'temp_pdf.pdf');
+        $repetidos = $this->repository->where([['hash_file', '=', $hash_file]])->all();
 
-        // $outputName = $paths["path_pdf"]."merged.pdf";
-        // // dd($outputName);
-        // $cmd = "gs -q -dNOPAUSE -dBATCH -sDEVICE=pdfwrite -sOutputFile=$outputName ";
-        // foreach($files as $file) {
-        //     $cmd .= $paths["path_pdf"]. $file." ";
-        // }
-        // $result = shell_exec($cmd);
+        if($repetidos->count() > 0 && $archivo->id_tipo_archivo != $repetidos[0]->id_tipo_archivo){
+            array_map( 'unlink', array_filter((array) glob($paths["path_pdf"] . '*') ) );
+            abort(403, 'El archivo ya ha sido registrado previamente como '.$repetidos[0]->ctgTipoArchivo->descripcion . ' de la empresa '.$archivo->empresa->razon_social ." (".$archivo->empresa->rfc.")");
+        }
 
-        // Storage::disk('padron_contratista')->put('PAT010101ABC/' .$archivo->ctgTipoArchivo->nombre.$archivo->complemento_nombre.'.'.$nombre_archivo[count($nombre_archivo)-1],  fopen($data['archivo'], 'r'))
         
-        dd($files);
-        dd('stop');
+
+        $nombre_archivo = explode('.', $data["archivo_nombre"]);
+        if(Storage::disk('padron_contratista')->put($directorio . '/' .$archivo->nombre_descarga.'.pdf', $pdf_file )){
+            $archivo->hash_file = $hash_file;
+            $archivo->nombre_archivo = $archivo->nombre_descarga;
+            $archivo->nombre_archivo_usuario = $data["archivo_nombre"];
+            $archivo->extension_archivo = 'pdf';
+            $archivo->save();
+            Storage::disk('padron_contratista')->put( 'hashfiles/' .$archivo->hash_file.'.pdf',  $paths["path_pdf"].'temp_pdf.pdf');
+        }else{
+            abort(403, 'Hubo un error al cargar el archivo, intente mas tarde');
+        }
         
+        $pdf = null;
+        fclose($pdf_file);
+        $this->removerCarpetas($paths["dir_pdf"]);
+
+        return $archivo;
+
+        
+    }
+
+    private function removerCarpetas($path){
+        gc_collect_cycles();
+        $di = new RecursiveDirectoryIterator($path, FilesystemIterator::SKIP_DOTS);
+        $ri = new RecursiveIteratorIterator($di, RecursiveIteratorIterator::CHILD_FIRST);
+        foreach ( $ri as $file ) {
+            $file->isDir() ?  rmdir($file) : @unlink($file);
+        }
+        return true;
+
     }
 
     private function generaDirectorios()
@@ -105,10 +143,10 @@ class ArchivoService
         $path_pdf = $dir_pdf . $nombre . "/";
         $path_zip = $dir_zip . $nombre_zip;
         if (!file_exists($dir_zip) && !is_dir($dir_zip)) {
-            mkdir($dir_zip, 777, true);
+            mkdir($dir_zip, 0777, true);
         }
         if (!file_exists($dir_pdf) && !is_dir($dir_pdf)) {
-            mkdir($dir_pdf, 777, true);
+            mkdir($dir_pdf, 0777, true);
         }
         return ["path_zip" => $path_zip, "path_pdf" => $path_pdf, "dir_pdf" => $dir_pdf];
     }
@@ -123,7 +161,7 @@ class ArchivoService
         }
         $storagePath  = Storage::disk('padron_contratista')->getDriver()->getAdapter()->getPathPrefix();
         // dd($storagePath . $directorio . '/' . $archivo->nombre_archivo);
-        return response()->file($storagePath . $directorio . '/' . $archivo->nombre_archivo );
+        return response()->file($storagePath . $directorio . '/' . $archivo->nombre_archivo . '.' . $archivo->extension_archivo );
     }
 
     public function getArchivosPrestadora($data){
